// src/types/commercetools.d.ts
export type AttributeDefinitionResult = {
	__typename?: 'AttributeDefinitionResult';
	limit?: Maybe<Scalars['Int']>;
	offset?: Maybe<Scalars['Int']>;
	total: Scalars['Int'];
	results: Array<AttributeDefinition>;
};
// ...
export type Initiator = {
	__typename?: 'Initiator';
	isPlatformClient?: Maybe<Scalars['Boolean']>;
	user?: Maybe<Reference>;
	externalUserId?: Maybe<Scalars['String']>;
	customer?: Maybe<Reference>;
	anonymousId?: Maybe<Scalars['String']>;
	clientId?: Maybe<Scalars['String']>;
};
// ...
export type LocalizedString = {
	__typename?: 'LocalizedString';
	locale: Scalars['Locale'];
	value: Scalars['String'];
};
// ...
export type Product = Versioned & {
	__typename?: 'Product';
	id: Scalars['String'];
	key?: Maybe<Scalars['String']>;
	version: Scalars['Long'];
	productTypeRef: Reference;
	productType?: Maybe<ProductTypeDefinition>;
	masterData: ProductCatalogData;
	catalogData?: Maybe<ProductCatalogData>;
	skus: Array<Scalars['String']>;
	createdAt: Scalars['DateTime'];
	lastModifiedAt: Scalars['DateTime'];
	stateRef?: Maybe<Reference>;
	state?: Maybe<State>;
	taxCategoryRef?: Maybe<Reference>;
	taxCategory?: Maybe<TaxCategory>;
	createdBy?: Maybe<Initiator>;
	lastModifiedBy?: Maybe<Initiator>;
};

export type ProductCatalogDataArgs = {
	id: Scalars['String'];
};

export type ProductAttributeInput = {
	name: Scalars['String'];
	value: Scalars['String'];
};

export type ProductCatalogData = {
	__typename?: 'ProductCatalogData';
	current?: Maybe<ProductData>;
	staged?: Maybe<ProductData>;
	published: Scalars['Boolean'];
	hasStagedChanges: Scalars['Boolean'];
};

export type ProductData = {
	__typename?: 'ProductData';
	name?: Maybe<Scalars['String']>;
	description?: Maybe<Scalars['String']>;
	nameAllLocales: Array<LocalizedString>;
	descriptionAllLocales?: Maybe<Array<LocalizedString>>;
	slug?: Maybe<Scalars['String']>;
	categoryOrderHint?: Maybe<Scalars['String']>;
	categoryOrderHints: Array<CategoryOrderHint>;
	categoriesRef: Array<Reference>;
	categories: Array<Category>;
	searchKeyword?: Maybe<Array<SearchKeyword>>;
	searchKeywords: Array<SearchKeywords>;
	metaTitle?: Maybe<Scalars['String']>;
	metaKeywords?: Maybe<Scalars['String']>;
	metaDescription?: Maybe<Scalars['String']>;
	masterVariant: ProductVariant;
	variants: Array<ProductVariant>;
	allVariants: Array<ProductVariant>;
	variant?: Maybe<ProductVariant>;
	skus: Array<Scalars['String']>;
};
// ...
export type ProductTypeDefinition = Versioned & {
	__typename?: 'ProductTypeDefinition';
	key?: Maybe<Scalars['String']>;
	name: Scalars['String'];
	description: Scalars['String'];
	attributeDefinitions: AttributeDefinitionResult;
	id: Scalars['String'];
	version: Scalars['Long'];
	createdAt: Scalars['DateTime'];
	lastModifiedAt: Scalars['DateTime'];
	createdBy?: Maybe<Initiator>;
	lastModifiedBy?: Maybe<Initiator>;
};
// ...
export type Reference = {
	__typename?: 'Reference';
	typeId: Scalars['String'];
	id: Scalars['String'];
};
// ...
export type State = Versioned & {
	__typename?: 'State';
	id: Scalars['String'];
	version: Scalars['Long'];
	key?: Maybe<Scalars['String']>;
	type: StateType;
	roles: Array<StateRole>;
	name?: Maybe<Scalars['String']>;
	nameAllLocales?: Maybe<Array<LocalizedString>>;
	description?: Maybe<Scalars['String']>;
	descriptionAllLocales?: Maybe<Array<LocalizedString>>;
	builtIn: Scalars['Boolean'];
	transitionsRef?: Maybe<Array<Reference>>;
	transitions?: Maybe<Array<State>>;
	initial: Scalars['Boolean'];
	createdAt: Scalars['DateTime'];
	lastModifiedAt: Scalars['DateTime'];
	createdBy?: Maybe<Initiator>;
	lastModifiedBy?: Maybe<Initiator>;
};
// ...
export enum StateRole {
	Return = 'Return',
	ReviewIncludedInStatistics = 'ReviewIncludedInStatistics',
}

export enum StateType {
	OrderState = 'OrderState',
	ProductState = 'ProductState',
	ReviewState = 'ReviewState',
	PaymentState = 'PaymentState',
	LineItemState = 'LineItemState',
}
// ...
export type TaxCategory = Versioned & {
	__typename?: 'TaxCategory';
	name: Scalars['String'];
	description?: Maybe<Scalars['String']>;
	rates: Array<TaxRate>;
	key?: Maybe<Scalars['String']>;
	id: Scalars['String'];
	version: Scalars['Long'];
	createdAt: Scalars['DateTime'];
	lastModifiedAt: Scalars['DateTime'];
	createdBy?: Maybe<Initiator>;
	lastModifiedBy?: Maybe<Initiator>;
};
// ...
export type TaxRate = {
	__typename?: 'TaxRate';
	name: Scalars['String'];
	amount: Scalars['Float'];
	includedInPrice: Scalars['Boolean'];
	country: Scalars['Country'];
	state?: Maybe<Scalars['String']>;
	id?: Maybe<Scalars['String']>;
	subRates: Array<SubRate>;
};
// ...
export type Versioned = {
	id: Scalars['String'];
	version: Scalars['Long'];
	createdAt: Scalars['DateTime'];
	lastModifiedAt: Scalars['DateTime'];
	createdBy?: Maybe<Initiator>;
	lastModifiedBy?: Maybe<Initiator>;
};